facturaapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pymupdf==1.24.10
pdfminer.six
pdfplumber==0.11.4
pandas==2.2.3